api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
anthropic>=0.18.0
pypdf2>=3.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
